# openapi.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AlexMubarakshin/ApiSchema/1.0.1
info:
  description: Api schema
  version: "1.0.1"
  title: API schema
security:
  - bearerAuth: []  
...
paths:
...
  /api/v1/post/{id}/comments:
    get:
      tags:
        - posts
      description: "Получение комментариев на публикации"
      summary: "Получение комментариев на публикации"
      parameters:
      - name: id
        in: path
        description: "ID публикации"
        required: true
        schema:
          type: number
          example: 1
      - $ref: '#/components/parameters/offsetParam'
      - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: "Успешное получение комментариев"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/Comment'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - posts
      description: "Создание комментария к публикации"
      summary: "Создание комментария к публикации"
      parameters:
      - name: id
        in: path
        description: "ID публикации"
        required: true
        schema:
          type: number
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentModel"
      responses:
        '200':
          description: "Успешное создание комментария"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: "#/components/schemas/Comment"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/post/{id}/comments/{comment_id}:
    put:
      tags:
        - posts
      description: "Редактирование комментария к публикации"
      summary: "Редактирование комментария к публикации"
      parameters:
      - name: id
        in: path
        description: "ID публикации"
        required: true
        schema:
          type: number
          example: 1
      - name: comment_id
        in: path
        description: "ID комментария публикации"
        required: true
        schema:
          type: number
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentModel"
      responses:
        '200':
          description: "Успешное создание комментария"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: "#/components/schemas/Comment"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - posts
      description: "Удаление комментария к публикации"
      summary: "Удаление комментария к публикации"
      parameters:
      - name: id
        in: path
        description: "ID публикации"
        required: true
        schema:
          type: number
          example: 1
      - name: comment_id
        in: path
        description: "ID комментария публикации"
        required: true
        schema:
          type: number
          example: 1
      responses:
        '200':
          description: "Успешное создание комментария"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        id:
                          type: number
                          description: "ID удаленной публикации"
                          example: 13
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/likes:
    get:
      tags:
        - likes
      description: "Получить список пользователей оставивших лайк"
      summary: "Получить список лайков"
      security:
        - bearerAuth: []
      parameters: 
        - name: item_id
          in: query
          required: true
          description: "ID объекта у которого необходимо получить \"Лайки\""
          schema:
            type: number
            example: 42
        - name: type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/LikeType"
      responses:
        "200":
          description: "Успешно"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: "#/components/schemas/LikeUsersList"

        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - likes
      description: "Поставить лайк"
      summary: "Поставить лайк"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeRequestBody"
      responses:
        "200":
          description: "Успешно"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: "#/components/schemas/LikeUsersList"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - likes
      description: "Убрать лайк"
      summary: "Убрать лайк"
      security:
        - bearerAuth: []
      parameters: 
        - name: item_id
          in: query
          required: true
          description: "ID объекта которому убирается \"Лайк\""
          schema:
            type: number
            example: 42
        - name: type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/LikeType"
      responses:
        "200":
          description: "Успешно"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        likes:
                          $ref: "#/components/schemas/LikesCount"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/friends:
    get:
      tags:
        - friends
      description: "Получить список друзей"
      summary: "Получить список друзей"
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          description: "Имя друга для поиска"
          required: false
          schema:
            type: string
            example: "Петр"
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: "Успешное получение списка друзей"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/Person'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/friends/{id}:
    delete:
      tags:
        - friends
      description: "Удалить пользователя из друзей"
      summary: "Удалить пользователя из друзей"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: "Id друга"
          required: true
          schema:
            type: string
            example: 1
      responses:
        "200":
          description: "Успешное удаление друга"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        message:
                          type: string
                          example: "ok"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - friends
      description: "Принять/Добавить пользователя в друзья"
      summary: "Принять/Добавить пользователя в друзья"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: "Id пользователя"
          required: true
          schema:
            type: string
            example: 1
      responses:
        "200":
          description: "Успешно"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        message:
                          type: string
                          example: "ok"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/friends/request:
    get:
      tags:
        - friends
      description: "Получить список входящик заявок на добавление в друзья"
      summary: "Получить список заявок"
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          description: "Имя пользователя для поиска по заявкам"
          required: false
          schema:
            type: string
            example: "Петр"
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: "Успешное получение списка заявок"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/Person'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/friends/recommendations:
    get:
      tags:
        - friends
      description: "Получить список рекомендаций"
      summary: "Получить список рекомендаций"
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: "Успешное получение списка рекомендаций"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/Person'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Тип ошибки"
          example: "invalid_request"
          enum:
            - invalid_request
            - unauthorized
        error_description:
          type: string
          items:
            $ref: '#/components/schemas/ErrorDescription'
    ErrorDescription:
      type: object
      properties:
        error_description:
          type: string
          description: "Error description"
          enum:
            - "Unauthorized"
            - "An authorization code must be supplied"
            - "Redirect URI mismatch"
            - "Invalid authorization code: CODE"
            - "Bad credentials"
    Response:
      type: object
      properties:
        error: 
          type: string
          description: "Сообщение, в случае ошибки здесь будет причина"
        timestamp:
          type: number
          example: 1559751301818
    ListResponse:
      allOf:
      - $ref: "#/components/schemas/Response"
      - type: object
        properties:
          total:
            type: integer
            description: "Количество элементов в списке"
            format: int64
          offset:
            type: integer
            description: "Отступ от начала списка"
            format: int64
          perPage:
            type: integer
            description: "Количество элементов на страницу"
            format: int64
            default: 20
...
    Person:
      type: object
      description: "Пользователь соц. сети"
      properties:
        id:
          type: number
          example: 1
        first_name:
          type: string
          example: "Петр"
        last_name:
          type: string
          example: "Петрович"
        reg_date:
          type: number
          example: 1559751301818
        birth_date:
          type: number
          example: 1559751301818
        email:
          type: string
          example: "petr@mail.ru"
        phone:
          type: string
          example: '89100000000'
        photo:
          type: string
          example: "https://...../photos/image123.jpg"
        about:
          type: string
          example: "Родился в небольшой, но честной семье"
        city:
          $ref: "#/components/schemas/City"
        country:
          $ref: "#/components/schemas/Country"
        messages_permission:
          type: string
          example: "ALL"
          description: "Разрешение на получение сообщений: ALL - от всех пользователей (кроме заблокированных), FRIENDS - только от друзей"
        last_online_time:
          type: number
          example: 1559751301818
        is_blocked:
          type: boolean
          example: false
...
    LikeRequestBody:
      type: object
      properties:
        item_id:
          type: number
          description: "ID объекта которому ставится \"Лайк\""
        type:
          $ref: "#/components/schemas/LikeType"
    LikeType:
      type: string
      description: "Тип объекта которому поставили лайк"
      example: "Post"
      enum:
        - Post
        - Comment
    LikesCount:
      type: string
      description: "Кол-во лайков"
      example: 1
    LikeUsersList:
      type: object
      properties:
        likes:
          $ref: "#/components/schemas/LikesCount"
        users:
          type: array
          description: "Пользователи поставившие лайк"
          items:
            type: string
            example: 1
...
    CommentModel:
      type: object
      description: "Комментарий к посту"
      properties:
        parent_id:
          type: number
          example: 1
          description: "Родительский комментарий (если ответ на комментарий к посту)"
        comment_text:
          type: string
    Comment:
      type: object
      description: "Комментарий к посту"
      allOf:
        - $ref: "#/components/schemas/CommentModel"
        - type: object
          properties:
            id:
              type: number
              description: "Id комментария"
              example: 111
            post_id:
              type: string
            time:
              type: number
              example: 1559751301818
            author_id:
              type: number
              example: 1
            is_blocked:
              type: boolean
              description: "Комментарий заблокирован"
...
    Country:
      type: object
      properties:
        id:
          type: number
          description: "Id страны"
          example: 1
        title:
          type: string
          description: "Название страны"
          example: "Россия"
    City:
      type: object
      properties:
        id:
          type: number
          description: "Id города"
          example: 1
        title:
          type: string
          description: "Название города"
          example: "Москва"
  parameters:
    offsetParam:
      in: query
      description: "Отступ от начала списка"
      name: offset
      required: false
      schema:
        type: integer
        format: int64
    limitParam:
      in: query
      description: "Количество элементов на страницу"
      name: itemPerPage
      required: false
      schema:
        default: 20
        type: integer
        format: int64